DjangoBackend/requirements.txt
Django>=5.2.6
djangorestframework>=3.16.0
django-cors-headers>=4.3.0

